# apps/orchestrator/internal/database/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"database/sql"
	"time"
)

type Deployment struct {
	ID           int64          `db:"id" json:"id"`
	ProjectID    int64          `db:"project_id" json:"project_id"`
	WorkerID     sql.NullInt64  `db:"worker_id" json:"worker_id"`
	PrNumber     int64          `db:"pr_number" json:"pr_number"`
	CommitSha    string         `db:"commit_sha" json:"commit_sha"`
	Branch       string         `db:"branch" json:"branch"`
	State        string         `db:"state" json:"state"`
	PreviewUrl   sql.NullString `db:"preview_url" json:"preview_url"`
	ContainerID  sql.NullString `db:"container_id" json:"container_id"`
	BuildLogPath sql.NullString `db:"build_log_path" json:"build_log_path"`
	ErrorMessage sql.NullString `db:"error_message" json:"error_message"`
	Attempt      int64          `db:"attempt" json:"attempt"`
	CreatedAt    sql.NullTime   `db:"created_at" json:"created_at"`
	UpdatedAt    sql.NullTime   `db:"updated_at" json:"updated_at"`
	StartedAt    sql.NullTime   `db:"started_at" json:"started_at"`
	CompletedAt  sql.NullTime   `db:"completed_at" json:"completed_at"`
}

type Domain struct {
	ID        int64        `db:"id" json:"id"`
	UserID    int64        `db:"user_id" json:"user_id"`
	Name      string       `db:"name" json:"name"`
	CreatedAt sql.NullTime `db:"created_at" json:"created_at"`
	UpdatedAt sql.NullTime `db:"updated_at" json:"updated_at"`
}

type Project struct {
	ID            int64        `db:"id" json:"id"`
	UserID        int64        `db:"user_id" json:"user_id"`
	Name          string       `db:"name" json:"name"`
	RepoUrl       string       `db:"repo_url" json:"repo_url"`
	VcsProvider   string       `db:"vcs_provider" json:"vcs_provider"`
	WebhookSecret string       `db:"webhook_secret" json:"webhook_secret"`
	CreatedAt     sql.NullTime `db:"created_at" json:"created_at"`
	UpdatedAt     sql.NullTime `db:"updated_at" json:"updated_at"`
}

type ProjectDomain struct {
	ID        int64        `db:"id" json:"id"`
	ProjectID int64        `db:"project_id" json:"project_id"`
	DomainID  int64        `db:"domain_id" json:"domain_id"`
	Prefix    string       `db:"prefix" json:"prefix"`
	CreatedAt sql.NullTime `db:"created_at" json:"created_at"`
}

type Session struct {
	ID        string       `db:"id" json:"id"`
	UserID    int64        `db:"user_id" json:"user_id"`
	Token     string       `db:"token" json:"token"`
	ExpiresAt time.Time    `db:"expires_at" json:"expires_at"`
	CreatedAt sql.NullTime `db:"created_at" json:"created_at"`
}

type User struct {
	ID           int64        `db:"id" json:"id"`
	Username     string       `db:"username" json:"username"`
	PasswordHash string       `db:"password_hash" json:"password_hash"`
	CreatedAt    sql.NullTime `db:"created_at" json:"created_at"`
	UpdatedAt    sql.NullTime `db:"updated_at" json:"updated_at"`
}

type Worker struct {
	ID           int64        `db:"id" json:"id"`
	Name         string       `db:"name" json:"name"`
	Address      string       `db:"address" json:"address"`
	State        string       `db:"state" json:"state"`
	Capacity     int64        `db:"capacity" json:"capacity"`
	CurrentLoad  int64        `db:"current_load" json:"current_load"`
	Capabilities string       `db:"capabilities" json:"capabilities"`
	LastSeenAt   sql.NullTime `db:"last_seen_at" json:"last_seen_at"`
	CreatedAt    sql.NullTime `db:"created_at" json:"created_at"`
	UpdatedAt    sql.NullTime `db:"updated_at" json:"updated_at"`
}